create-dmg: update bindata
The change in https://github.com/mitchellh/gon/pull/73 alone was not sufficient
to solve https://github.com/mitchellh/gon/issues/64 since it did not include an
update to the embedded bindata for create-dmg. So this commit rebuilds the
bindata in order to make the change take effect.

Ran the following:

  # install go-bindata
  # as per https://github.com/go-bindata/go-bindata/issues/77#issuecomment-931578849
  go install -a -v github.com/go-bindata/go-bindata/...@latest

  # run go-bindata to update create-dmg from vendor dir
  # as per https://github.com/UniversalMediaServer/gon/blob/master/build.sh#L3
  cd internal/createdmg/bindata
  go-bindata -prefix '../../../vendor/create-dmg' -pkg bindata ../../../vendor/create-dmg/*